requirements.txt
colorama==0.4.3
colour==0.1.5
pyserial==3.5
style==1.1.0
mongita==1.1.1
paho-mqtt==1.6.1
pynmeagps==1.0.10
adafruit-circuitpython-bno08x==1.1.3
requests==2.28.1
calypso-anemometer==0.6.0
aioretry==5.0.2